requirements.txt
flask==3.0.2
flash_attn==2.7.4.post1
torch==2.6.0
transformers==4.48.2
accelerate==1.3.0
soundfile==0.13.1
pillow>=7.1.0
scipy==1.15.2
torchvision
backoff==2.2.1
peft==0.13.2
streamlit==1.31.1
requests==2.31.0
audio-recorder-streamlit==0.0.8
numpy==1.26.4